fixed all da bugs

+++ lambdas/update_supporter_settings.py
@@ -21,62 +21,69 @@ def update_supporter_settings(event, context):
 
     #Check if supporter exists
     sql = 'SELECT * FROM supporters WHERE supporter_id = :supporter_id;'
-    supporter_id_param = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}]
-    response = query(sql, supporter_id_param)
+    supporter_id_param = {'name': 'supporter_id', 'value': {'longValue': supporter_id}}
+    response = query(sql, [supporter_id_param])
     if(response['records'] == []):
         raise LambdaException("404: Supporter does not exist")
 
-
-    link_type_list = event['links']
     link_id_list = []
-    #Check if link_type exists, and convert to link_id
-    for link_type, link in link_type_list:
-        sql = 'SELECT link_id FROM link WHERE link_type = :link_type;'
-        sql_parameters = [{'name': 'link_type', 'value': {'longValue': link_type}}]
-        response = query(sql, sql_parameters)
-        if(response['records'] ==[]):
-            raise LambdaException("404: link_id:" + str(entry) + " does not exist")
-        else:
-            link_id_list.append(response['records'][0][0]['longValue'])
+    if 'links' in event:
+        link_type_list = event['links']
+        
+        #Check if link_type exists, and convert to link_id
+        for link_type, link in link_type_list:
+            sql = 'SELECT link_id FROM link WHERE link_type = :link_type;'
+            sql_parameters = [{'name': 'link_type', 'value': {'longValue': link_type}}]
+            response = query(sql, sql_parameters)
+            if(response['records'] ==[]):
+                raise LambdaException("404: link_id:" + str(entry) + " does not exist")
+            else:
+                link_id_list.append(response['records'][0][0]['longValue'])
 
 
-    tag_type_list = event['tags']
     tag_type_id_list = []
-    #Check if tag_type exists, and convert to tag_type_id
-    for entry in tag_type_list:
-        sql = 'SELECT tag_type_id FROM tag_type WHERE tag_type = :tag_type;'
-        sql_parameters = [{'name': 'tag_type', 'value': {'longValue': entry}}]
-        response = query(sql, sql_parameters)
-        if(response['records'] ==[]):
-            raise LambdaException("404: tag_type_id:" + str(entry) + " does not exist")
-        else:
-            tag_type_id_list.append(response['records'][0][0]['longValue'])
+    if 'tags' in event:
+        tag_type_list = event['tags']
+        
+        #Check if tag_type exists, and convert to tag_type_id
+        for entry in tag_type_list:
+            sql = 'SELECT tag_type_id FROM tag_type WHERE tag_type = :tag_type;'
+            sql_parameters = [{'name': 'tag_type', 'value': {'longValue': entry}}]
+            response = query(sql, sql_parameters)
+            if(response['records'] ==[]):
+                raise LambdaException("404: tag_type_id:" + str(entry) + " does not exist")
+            else:
+                tag_type_id_list.append(response['records'][0][0]['longValue'])
 
 
-    major_list = event['major_preferences']
     major_id_list = []
-    #Check if major_id exists
-    for entry in major_list:
-        sql = 'SELECT major_id FROM major WHERE major = :major;'
-        sql_parameters = [{'name': 'major', 'value': {'longValue': entry}}]
-        response = query(sql, sql_parameters)
-        if(response['records'] ==[]):
-            raise LambdaException("404: major_id:" + str(entry) + " does not exist")
-        else:
-            major_id_list.append(response['records'][0][0]['longValue'])
+    if 'major_preferences' in event:
+        major_list = event['major_preferences']
+        
+        #Check if major_id exists
+        for entry in major_list:
+            sql = 'SELECT major_id FROM major WHERE major = :major;'
+            sql_parameters = [{'name': 'major', 'value': {'longValue': entry}}]
+            response = query(sql, sql_parameters)
+            if(response['records'] ==[]):
+                raise LambdaException("404: major_id:" + str(entry) + " does not exist")
+            else:
+                major_id_list.append(response['records'][0][0]['longValue'])
 
     
-    #Check if minor_id exists
-    minor_list = event['minor_preferences']
     minor_id_list = []
-    for entry in minor_list:
-        sql = 'SELECT minor_id FROM minor WHERE minor = :minor;'
-        sql_parameters = [{'name': 'minor', 'value': {'longValue': entry}}]
-        response = query(sql, sql_parameters)
-        if(response['records'] ==[]):
-            raise LambdaException("404: minor_id:" + str(entry) + " does not exist")
-        else:
-            minor_id_list.append(response['records'][0][0]['longValue'])
+    if 'minor_preferences' in event:
+        minor_list = event['minor_preferences']
+        
+        #Check if minor_id exists
+        for entry in minor_list:
+            sql = 'SELECT minor_id FROM minor WHERE minor = :minor;'
+            sql_parameters = [{'name': 'minor', 'value': {'longValue': entry}}]
+            response = query(sql, sql_parameters)
+            if(response['records'] ==[]):
+                raise LambdaException("404: minor_id:" + str(entry) + " does not exist")
+            else:
+                minor_id_list.append(response['records'][0][0]['longValue'])
 
 
     #User settings
@@ -108,17 +115,23 @@ def update_supporter_settings(event, context):
         user_settings['bio'] = bio
         updated_user_vals.append("bio = :bio")
 
-    if event['pronouns'] ! = "":
+    if event['pronouns'] != "":
         pronouns = event['pronouns']
         user_settings['pronouns'] = pronouns
         updated_user_vals.append("pronouns = :pronouns")
+        
+    if event['gender'] != "":
+        gender = event['gender']
+        user_settings['gender'] = gender
+        updated_user_vals.append("gender = :gender")
 
     if event['phone'] != "":
         phone = event['phone']
         user_settings['phone'] = phone
         updated_user_vals.append("phone = :phone")
 
-    user_settings_sql = "UPDATE users SET" + ", ".join(updated_user_vals) + " WHERE user_id = :supporter_id;"
+    user_settings_sql = "UPDATE users SET " + ", ".join(updated_user_vals) + " WHERE id = :supporter_id;"
+    print(user_settings_sql)
     user_settings_params = dictionary_to_list(user_settings) 
     user_settings_params.append(supporter_id_param)
 
@@ -153,7 +166,7 @@ def update_supporter_settings(event, context):
         supporter_settings['office'] = office
         updated_supporter_vals.append("office = :office")
 
-    supporter_settings_sql = "UPDATE supporters SET" + ", ".join(updated_supporter_vals) + " WHERE supporter_id = :supporter_id;"
+    supporter_settings_sql = "UPDATE supporters SET " + ", ".join(updated_supporter_vals) + " WHERE supporter_id = :supporter_id;"
     supporter_settings_params = dictionary_to_list(supporter_settings)
     supporter_settings_params.append(supporter_id_param)
 
@@ -161,7 +174,7 @@ def update_supporter_settings(event, context):
         try:
             query(supporter_settings_sql, supporter_settings_params)
         except Exception as e:
-            raise LambdaException("500: Unable to update supporter settings, ")
+            raise LambdaException("500: Unable to update supporter settings, " + str(e))
 
 
     #Supporter Specializations
@@ -169,21 +182,35 @@ def update_supporter_settings(event, context):
     updated_specializations = []
 
     if event['specialization_type'] != "":
-        specialization_type_id = event['specialization_type_id']
+        specialization_type = event['specialization_type']
+        specialization_type_param = [{'name' : 'specialization_type', 'value' : {'stringValue' : specialization_type}}]
+        
+        spec_type_id_sql = "SELECT specialization_type_id FROM specialization_type WHERE specialization_type = :specialization_type;"
+        specialization_type_id = ""
+        try:
+            specialization_type_id = query(spec_type_id_sql, specialization_type_param)['records']
+        except Exception as e:
+            raise LambdaException("500: Unable to get specialization_type_id")
+            
+        if specialization_type_id != []:
+            specialization_type_id = specialization_type_id[0][0]['longValue']
+        else:
+            raise LambdaException("404: Specialization type does not exist")
+        
         supporter_specialization['specialization_type_id'] = specialization_type_id
         updated_specializations.append("specialization_type_id = :specialization_type_id")
 
     if event['max_students'] != "":
-        max_students = event['max_students']
+        max_students = int(event['max_students'])
         supporter_specialization['max_students'] = max_students
         updated_specializations.append("max_students = :max_students")
 
     if event['duration'] != "":
-        duration = event['duration']
+        duration = int(event['duration'])
         supporter_specialization['duration'] = duration
         updated_specializations.append("duration = :duration")
 
-    specializations_sql = "UPDATE supporter_specializations SET" + ", ".join(updated_specializations) + " WHERE supporter_id = :supporter_id;"
+    specializations_sql = "UPDATE supporter_specializations SET " + ", ".join(updated_specializations) + " WHERE supporter_id = :supporter_id;"
     specializations_params = dictionary_to_list(supporter_specialization)
     specializations_params.append(supporter_id_param)
 
@@ -208,7 +235,7 @@ def update_supporter_settings(event, context):
         supporter_student_prefs['hours_before_appointment'] = hours_before_appointment
         updated_student_prefs.append("hours_before_appointment = :hours_before_appointment")
 
-    student_prefs_sql = "UPDATE supporter_preferences_for_students SET" + ", ".join(updated_student_prefs) + " WHERE supporter_id = :supporter_id;"
+    student_prefs_sql = "UPDATE supporter_preferences_for_students SET " + ", ".join(updated_student_prefs) + " WHERE supporter_id = :supporter_id;"
     student_prefs_params = dictionary_to_list(supporter_student_prefs)
     student_prefs_params.append(supporter_id_param)
 
@@ -221,19 +248,20 @@ def update_supporter_settings(event, context):
     
     #User links
     #Execute parameterized query to delete supporter's old links
-    sql = "DELETE FROM user_links WHERE use_id = :supporter_id;"
-    try:
-        query(sql, supporter_id_param)
-    except Exception as e:
-        raise LambdaException("500: Unable to delete links")
-
-    #Execute parameterized queries for updating links
-    for link_id, link in link_id_list:
-        sql = 'INSERT INTO user_links(user_id, link_id, link) VALUES (:user_id, :link_id, :link);'
-        sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'link_id', 'value': {'longValue': link_id}}, {'name': 'link', 'value': {'longValue': link}}]
-        major_response = query(sql, sql_parameters)
-        if(major_response["numberOfRecordsUpdated"] == 0):
-            raise LambdaException("409: User links not updated")
+    if len(link_id_list) > 0:
+        sql = "DELETE FROM user_link WHERE user_id = :supporter_id;"
+        try:
+            query(sql, [supporter_id_param])
+        except Exception as e:
+            raise LambdaException("500: Unable to delete links" + str(e))
+    
+        #Execute parameterized queries for updating links
+        for link_id, link in link_id_list:
+            sql = 'INSERT INTO user_links(user_id, link_id, link) VALUES (:user_id, :link_id, :link);'
+            sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'link_id', 'value': {'longValue': link_id}}, {'name': 'link', 'value': {'longValue': link}}]
+            major_response = query(sql, sql_parameters)
+            if(major_response["numberOfRecordsUpdated"] == 0):
+                raise LambdaException("409: User links not updated")
 
 
 
@@ -246,50 +274,53 @@ def update_supporter_settings(event, context):
 
     #Supporter Tags
     #Execute parameterized query to delete supporter's old tags
-    sql = "DELETE FROM supporter_tags WHERE supporter_id = :supporter_id;"
-    try:
-        query(sql, supporter_id_param)
-    except Exception as e:
-        raise LambdaException("500: Unable to delete tags")
-
-    #Execute parameterized queries for updating tags
-    for entry in tag_type_id_list:
-        sql = 'INSERT INTO supporter_tags(supporter_id, tag_type) VALUES (:supporter_id, :tag_type_id);'
-        sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'tag_type_id', 'value': {'longValue': entry}}]
-        major_response = query(sql, sql_parameters)
-        if(major_response["numberOfRecordsUpdated"] == 0):
-            raise LambdaException("409: supporter_tags majors not updated")
+    if len('tag_type_id_list') > 0:
+        sql = "DELETE FROM supporter_tags WHERE supporter_id = :supporter_id;"
+        try:
+            query(sql, [supporter_id_param])
+        except Exception as e:
+            raise LambdaException("500: Unable to delete tags")
+    
+        #Execute parameterized queries for updating tags
+        for entry in tag_type_id_list:
+            sql = 'INSERT INTO supporter_tags(supporter_id, tag_type) VALUES (:supporter_id, :tag_type_id);'
+            sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'tag_type_id', 'value': {'longValue': entry}}]
+            major_response = query(sql, sql_parameters)
+            if(major_response["numberOfRecordsUpdated"] == 0):
+                raise LambdaException("409: supporter_tags majors not updated")
     
 
     #Execute parameterized query to delete supporter's old major preferences
-    sql = 'DELETE FROM supporter_major_preferences WHERE supporter_id = :supporter_id;'
-    try:
-        query(sql, supporter_id_param)
-    except Exception as e:
-        raise LambdaException("500: Unable to delete major preferences, " + str(e))
-
-    #Execute parameterized queries for updating major preferences
-    for entry in major_id_list:
-        sql = 'INSERT INTO supporter_major_preferences(supporter_id, major_id) VALUES (:supporter_id, :major_id);'
-        sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'major_id', 'value': {'longValue': entry}}]
-        major_response = query(sql, sql_parameters)
-        if(major_response["numberOfRecordsUpdated"] == 0):
-            raise LambdaException("409: supporter_major_preferences majors not updated")
+    if len(major_id_list) > 0:
+        sql = 'DELETE FROM supporter_major_preferences WHERE supporter_id = :supporter_id;'
+        try:
+            query(sql, [supporter_id_param])
+        except Exception as e:
+            raise LambdaException("500: Unable to delete major preferences, " + str(e))
+    
+        #Execute parameterized queries for updating major preferences
+        for entry in major_id_list:
+            sql = 'INSERT INTO supporter_major_preferences(supporter_id, major_id) VALUES (:supporter_id, :major_id);'
+            sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'major_id', 'value': {'longValue': entry}}]
+            major_response = query(sql, sql_parameters)
+            if(major_response["numberOfRecordsUpdated"] == 0):
+                raise LambdaException("409: supporter_major_preferences majors not updated")
 
 
     #Execute parameterized query to delete supporter's old minor preferences
-    sql = 'DELETE FROM supporter_minor_preferences WHERE supporter_id = :supporter_id;'
-    try:
-        query(sql, supporter_id_param)
-    except Exception as e:
-        raise LambdaException("500: Unable to delete minor preferences")
-
-    for entry in minor_id_list:
-        sql = 'INSERT INTO supporter_major_preferences(supporter_id, minor_id) VALUES (:supporter_id, :minor_id);'
-        sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'minor_id', 'value': {'longValue': entry}}]
-        minor_response = query(sql, sql_parameters)
-        if(minor_response["numberOfRecordsUpdated"] == 0):
-            raise LambdaException("409: supporter_minorr_preferences minors not updated")
+    if len(minor_id_list) > 0:
+        sql = 'DELETE FROM supporter_minor_preferences WHERE supporter_id = :supporter_id;'
+        try:
+            query(sql, [supporter_id_param])
+        except Exception as e:
+            raise LambdaException("500: Unable to delete minor preferences")
+    
+        for entry in minor_id_list:
+            sql = 'INSERT INTO supporter_major_preferences(supporter_id, minor_id) VALUES (:supporter_id, :minor_id);'
+            sql_parameters = [{'name': 'supporter_id', 'value': {'longValue': supporter_id}}, {'name': 'minor_id', 'value': {'longValue': entry}}]
+            minor_response = query(sql, sql_parameters)
+            if(minor_response["numberOfRecordsUpdated"] == 0):
+                raise LambdaException("409: supporter_minorr_preferences minors not updated")
 
 
     #Update supporter types
@@ -315,7 +346,7 @@ def update_supporter_settings(event, context):
 
         supporter_types_sql = "UPDATE supporter_type SET " + ", ".join(updated_supporter_types) + " WHERE supporter_id = :supporter_id;"
         try:
-            query(supporter_types_sql, supporter_id_param)
+            query(supporter_types_sql, [supporter_id_param])
         except Exception as e:
             raise LambdaException("500: Unable to update supporter_types, " + str(e))
